additional scripts

+++ aliases/kill-port.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
+if [ "$1" ]
+  then
+    { #try 
+        kill -9 $(lsof -t -i:$1) 2>/dev/null
+    } || { # catch
+        script=$HOME/bashrc/aliases/batch/kill-port.bat
+        cmd "/C $($HOME/bashrc/aliases/convert-path.sh $script) $1"
+    }
+fi
 
-{ #try 
-    kill -9 $(lsof -t -i:$1) 2>/dev/null
-} || { # catch
-    ECHO hello world
-    batFile="$PWD"/kill-port.bat
-    ECHO FOR \/F \"tokens=4 delims= \" \%\%P IN \(\'netstat -a -n -o \^\| findstr :$1\'\) DO TaskKill.exe /PID \%\%P > $batFile
-    rm $batFile 
-}
